requirements.txt
selenium==4.15.0
python-telegram-bot==20.6
